Move user form page to components.
