+++ src/components/index.ts
@@ -37,3 +37,4 @@ export {
 export { APIButton } from './api-button/api-button';
 export { Main } from './patternfly-wrappers/main';
 export { UserForm } from './user-form/user-form';
+export { UserFormPage } from './user-form/user-form-page';
