+++ src/utilities/index.ts
@@ -1,3 +1,4 @@
 export { convertContentSummaryCounts } from './content-summary';
 export { ParamHelper } from './param-helper';
 export { sanitizeDocsUrls } from './sanitize-docs-urls';
+export { mapErrorMessages } from './map-error-messages';
